Fix GCC type conversion complain
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@11845 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/algo/align/demo/splign/util.cpp
@@ -173,7 +173,7 @@ void MakeIDX()
   string s;
   while((c = cin.get()) && c != EOF) {
     if(c == '>') {
-      pos = cin.tellg() - 1;
+      pos = size_t(cin.tellg()) - 1;
       cin >> s;
       cout << s << '\t' << pos << endl;
     }
